Fix error below:
RuntimeError: Expected tensor for argument #1 'indices' to have scalar type Long; but got torch.cuda.FloatTensor instead (while checking arguments for embedding)

+++ openspeech/data/text/data_loader.py
@@ -45,8 +45,8 @@ def _collate_fn(batch, pad_id: int = 0):
 
     batch_size = len(batch)
 
-    inputs = torch.zeros(batch_size, max_seq_size).fill_(pad_id)
-    targets = torch.zeros(batch_size, max_seq_size).fill_(pad_id)
+    inputs = torch.zeros(batch_size, max_seq_size).fill_(pad_id).long()
+    targets = torch.zeros(batch_size, max_seq_size).fill_(pad_id).long()
 
     for x in range(batch_size):
         sample = batch[x]
